% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_function.R
\name{augmented_est}
\alias{augmented_est}
\title{Main augmented estimate function for extendAUG}
\usage{
augmented_est(m, x, s, y, val_x, val_s, val_y)
}
\arguments{
\item{m}{size of validation set (i.e., 100, 200, 400, & 800)}

\item{x}{n*i matrix (predictors with dim n: size of full data times i: num of predictors)}

\item{s}{surrogate phenotypes}

\item{y}{true phenotypes (with NA)}

\item{val_x}{value of x}

\item{val_s}{value of s}

\item{val_y}{value of y}
}
\value{
estimates, variance, and ci of three models (M1 valid, M2 native & M4 augment)
}
\description{
Main augmented estimate function for extendAUG
}
