% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Data generator function for extendAUG}
\usage{
data_generator(n = 2000, m = 100)
}
\arguments{
\item{n}{size of full data}

\item{m}{size of validation set}
}
\value{
validation matrix of x, y, s, and their values (for augmented_est function)
}
\description{
Data generator function for extendAUG
}
