% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo_AugEst.R
\name{homo_AugEst}
\alias{homo_AugEst}
\title{Run extendAUG simulation (Replication and Extension of the Augmented Estimation)}
\usage{
homo_AugEst(n = 2000, Nsim = 10, x_beta = 2, parallel_ifrun = FALSE)
}
\arguments{
\item{n}{size of full data: integer, greater than 1000 (i.e., 2,000, 5,000, 10,000 )}

\item{Nsim}{number of iterations: integer}

\item{x_beta}{coefficients of the covariates ("1" or "2":  1 -- coefficient of x1; 2 -- coefficient of x2)}

\item{parallel_ifrun}{boolen condition}
}
\value{
Augmented estimator of three models for four size of validation set (m = 100, 200, 400, & 800)
}
\description{
Run extendAUG simulation (Replication and Extension of the Augmented Estimation)
}
