% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_time_simulation.R
\name{one_time_simulation}
\alias{one_time_simulation}
\title{Run one-time simulation for extendAUG}
\usage{
one_time_simulation(Nsim = 1, n = 2000, x_beta = 2)
}
\arguments{
\item{Nsim}{number of iterations: integer,1}

\item{n}{size of full data: integer (i.e., 2,000,5,000,10,000)}

\item{x_beta}{coefficients of the covariates ("1" or "2":  1 -- coefficient of x1; 2 -- coefficient of x2)}
}
\value{
Augmented estimator of three models for four size of validation set (m = 100,200, 400, &800)
}
\description{
Run one-time simulation for extendAUG
}
